LPS-122233 No need of this interface
